move more to operations dir
