+++ lib/core/traits.rs
@@ -11,8 +11,8 @@ pub trait Framework {
     fn detect(&self, target: &Target) -> bool;
     fn runnables(&self, target: &Target) -> Result<Vec<Runnable>, FrameworkError>;
     fn generate_command(&self, runnable: Runnable) -> Command;
-    fn capabilities(&self) -> &HashSet<CapabilityDetails>;
-    fn search_for_capability(&self, description: &str) -> Option<&CapabilityDetails>;
+    fn capabilities(&self) -> HashSet<CapabilityDetails>;
+    fn search_for_capability(&self, description: &str) -> Option<CapabilityDetails>;
 }
 
 pub trait FrameworkProvider {
