+++ lib/framework/golang/gotest.rs
@@ -1,11 +1,8 @@
 use std::collections::HashSet;
-use std::ops::Range;
-use std::sync::LazyLock;
 
+use crate::core::enums::Language as crate_language;
 use crate::core::errors::FrameworkError;
-use crate::core::metadata::RunnableMeta;
 use crate::core::types::Command;
-use crate::core::types::CursorPosition;
 use crate::core::types::Runnable;
 use crate::core::types::Target;
 use crate::core::{
@@ -13,39 +10,40 @@ use crate::core::{
     traits::{Framework, FrameworkProvider},
     types::CapabilityDetails,
 };
-use crate::framework::golang::treesitter::operations::extract_gotest_subtests;
-use crate::framework::golang::treesitter::queries::gotest_file_test_methods;
-use crate::framework::golang::treesitter::queries::gotest_test_function;
-
-use crate::core::enums::Language as crate_language;
+use crate::framework::golang::treesitter::operations::gotest_get_file_tests;
+use crate::framework::golang::treesitter::operations::gotest_get_subtests;
+use crate::framework::golang::treesitter::operations::gotest_get_test;
 use crate::framework::golang::treesitter::operations::parse_tree;
-use crate::treesitter::node as crate_treesitter_node;
-use tree_sitter::{Language, Node, Query, QueryCursor};
 
-pub struct GotestProvider;
+pub struct GotestProvider {
+    search_capabilities: HashSet<CapabilityDetails>,
+}
 
 static FILE_SUFFIX: &str = "_test.go";
 
-static SEARCH_STRATEGIES: LazyLock<HashSet<CapabilityDetails>> = LazyLock::new(|| {
-    let mut res = HashSet::with_capacity(3);
-    res.insert(CapabilityDetails {
-        capability: Capability::TestRunner,
-        search: crate::core::enums::Search::Nearest,
-        description: "Test Nearest".to_string(),
-    });
-    res.insert(CapabilityDetails {
-        capability: Capability::TestRunner,
-        search: crate::core::enums::Search::Method,
-        description: "Test Function".to_string(),
-    });
-    res.insert(CapabilityDetails {
-        capability: Capability::TestRunner,
-        search: crate::core::enums::Search::File,
-        description: "Test File".to_string(),
-    });
-
-    res
-});
+impl GotestProvider {
+    pub fn new() -> Self {
+        let mut res = HashSet::with_capacity(3);
+        res.insert(CapabilityDetails {
+            capability: Capability::TestRunner,
+            search: crate::core::enums::Search::Nearest,
+            description: "Test Nearest".to_string(),
+        });
+        res.insert(CapabilityDetails {
+            capability: Capability::TestRunner,
+            search: crate::core::enums::Search::Method,
+            description: "Test Function".to_string(),
+        });
+        res.insert(CapabilityDetails {
+            capability: Capability::TestRunner,
+            search: crate::core::enums::Search::File,
+            description: "Test File".to_string(),
+        });
+        Self {
+            search_capabilities: res,
+        }
+    }
+}
 
 impl FrameworkProvider for GotestProvider {
     fn create(&self) -> Box<dyn Framework> {
@@ -105,7 +103,7 @@ impl Framework for GotestProvider {
         let walker_node = walker.node();
         match target.search_strategy {
             crate::core::enums::Search::File => {
-                let parent_runnables = self.get_all_test_methods(walker_node, target);
+                let parent_runnables = gotest_get_file_tests::op::execute(walker_node, target);
                 if parent_runnables.is_none() {
                     return Err(FrameworkError::NotFoundError(
                         "Go Test Function not found no tests in this file".to_string(),
@@ -114,11 +112,8 @@ impl Framework for GotestProvider {
                 let parent_runnables = parent_runnables.unwrap();
                 let mut res: Vec<Runnable> = vec![];
                 for parent in parent_runnables.into_iter() {
-                    let subtests = extract_gotest_subtests::op::execute(
-                        walker_node,
-                        parent.to_owned(),
-                        target,
-                    );
+                    let subtests =
+                        gotest_get_subtests::op::execute(walker_node, parent.to_owned(), target);
                     if let Some(sub) = subtests {
                         res.extend(sub);
                     } else {
@@ -128,7 +123,7 @@ impl Framework for GotestProvider {
                 Ok(res)
             }
             crate::core::enums::Search::Method => {
-                let res = self.get_single_test_method(walker_node, target);
+                let res = gotest_get_test::op::execute(walker_node, target);
                 if res.is_none() {
                     return Err(FrameworkError::NotFoundError(
                         "Go Test Function not found at position".to_string(),
@@ -137,7 +132,7 @@ impl Framework for GotestProvider {
                 Ok(vec![res.unwrap()])
             }
             crate::core::enums::Search::Nearest => {
-                let parent_runnable = self.get_single_test_method(walker_node, target);
+                let parent_runnable = gotest_get_test::op::execute(walker_node, target);
                 if parent_runnable.is_none() {
                     return Err(FrameworkError::NotFoundError(
                         "Go Test Function not found at position".to_string(),
@@ -145,7 +140,7 @@ impl Framework for GotestProvider {
                 }
 
                 let parent_runnable = parent_runnable.unwrap();
-                let subtests = extract_gotest_subtests::op::execute(
+                let subtests = gotest_get_subtests::op::execute(
                     walker_node,
                     parent_runnable.to_owned(),
                     target,
@@ -159,118 +154,15 @@ impl Framework for GotestProvider {
         }
     }
 
-    fn capabilities(&self) -> &HashSet<CapabilityDetails> {
-        LazyLock::force(&SEARCH_STRATEGIES)
+    fn capabilities(&self) -> HashSet<CapabilityDetails> {
+        self.search_capabilities.clone()
     }
 
-    fn search_for_capability(&self, description: &str) -> Option<&CapabilityDetails> {
-        let c = LazyLock::force(&SEARCH_STRATEGIES);
-        c.iter().find(|s| s.description == description).clone()
-    }
-}
-
-impl GotestProvider {
-    pub fn new() -> Self {
-        Self {}
-    }
-
-    fn get_test_function_query(&self) -> Option<Query> {
-        let query_pattern = gotest_test_function::query();
-        let query = Query::new(&Language::new(tree_sitter_go::LANGUAGE), &query_pattern).ok()?;
-        Some(query)
-    }
-
-    fn get_single_test_method(&self, node: Node, target: &Target) -> Option<Runnable> {
-        let current_node_position = node.start_position();
-        let query = self.get_test_function_query()?;
-        let content = target.buffer.content;
-        let test_name_index = query.capture_index_for_name("test_name")?;
-        let test_function_index = query.capture_index_for_name("testfunc")?;
-        let mut cursor = QueryCursor::new();
-        let query_matches = cursor.matches(&query, node, content.as_bytes());
-        for node_matched in query_matches.into_iter() {
-            let function_node = node_matched
-                .captures
-                .iter()
-                .filter(|c| c.index == test_function_index)
-                .map(|c| c.node)
-                .next();
-
-            if function_node.is_none() {
-                continue;
-            }
-
-            let function_node = function_node.unwrap();
-
-            for m in node_matched
-                .captures
-                .iter()
-                .filter(|c| c.index == test_name_index)
-            {
-                if m.node.start_position().row <= current_node_position.row
-                    && m.node.end_position().row >= current_node_position.row
-                {
-                    return Some(Runnable {
-                        name: crate_treesitter_node::node_text(m.node, content),
-                        filepath: target.buffer.filepath.to_string(),
-                        range: Range {
-                            start: CursorPosition::from_point(function_node.start_position()),
-                            end: CursorPosition::from_point(function_node.end_position()),
-                        },
-                        meta: RunnableMeta::default_golang(),
-                    });
-                }
-            }
-        }
-
-        None
-    }
-
-    fn get_all_test_methods(&self, node: Node, target: &Target) -> Option<Vec<Runnable>> {
-        let content = target.buffer.content;
-        let query_pattern = &gotest_file_test_methods::query();
-        let query = Query::new(&Language::new(tree_sitter_go::LANGUAGE), query_pattern).ok()?;
-        let test_name_index = query.capture_index_for_name("test_name")?;
-        let test_function_index = query.capture_index_for_name("testfunc")?;
-        let mut cursor = QueryCursor::new();
-        let query_matches = cursor.matches(&query, node, content.as_bytes());
-
-        let mut parent_runnables: Vec<Runnable> = vec![];
-        for node_matched in query_matches {
-            let function_node = node_matched
-                .captures
-                .iter()
-                .filter(|c| c.index == test_function_index)
-                .map(|c| c.node)
-                .next();
-
-            if function_node.is_none() {
-                continue;
-            }
-
-            let function_node = function_node.unwrap();
-
-            for m in node_matched
-                .captures
-                .iter()
-                .filter(|c| c.index == test_name_index)
-            {
-                parent_runnables.push(Runnable {
-                    name: crate_treesitter_node::node_text(m.node, content),
-                    filepath: target.buffer.filepath.to_string(),
-                    range: Range {
-                        start: CursorPosition::from_point(function_node.start_position()),
-                        end: CursorPosition::from_point(function_node.end_position()),
-                    },
-                    meta: RunnableMeta::default_golang(),
-                });
-            }
-        }
-
-        if parent_runnables.is_empty() {
-            None
-        } else {
-            Some(parent_runnables)
-        }
+    fn search_for_capability(&self, description: &str) -> Option<CapabilityDetails> {
+        let capabilities = self.capabilities();
+        capabilities
+            .iter()
+            .find(|&s| s.description == description)
+            .cloned()
     }
 }
