+++ lib/framework/golang/treesitter/operations/mod.rs
@@ -1,3 +1,5 @@
-pub(crate) mod extract_gotest_subtests;
 pub(crate) mod get_build_tags;
+pub(crate) mod gotest_get_file_tests;
+pub(crate) mod gotest_get_subtests;
+pub(crate) mod gotest_get_test;
 pub(crate) mod parse_tree;
